requirements.txt
aiohttp==3.9.5
aiohttp-apispec==1.5.0
aiohttp-cors==0.7.0
aiohttp-jinja2==1.5.0
aiohttp-security==0.4.0
aiohttp-session==2.8.0
aiodns==2.0.0
aiopg==1.4.0
aioredis==1.3.1
anybadge==1.11.1
sqlalchemy==1.3.16
alembic==1.4.2
aiocache[redis]==0.11.1
webargs==5.5.3
pydantic==2.7.3
pydantic-settings==2.3.1

pytest-cov==2.8.1
pytest-aiohttp==0.3.0
pytest-mock==3.1.0
pytest-freezegun==0.4.2
factory_boy==3.2.0

pydocstyle==3.0.0
flake8==3.9.0
flake8-blind-except==0.1.1
flake8-bugbear==19.3.0
flake8-builtins==1.4.1
flake8-commas==2.0.0
flake8-comprehensions==2.1.0
flake8-debugger==3.1.0
flake8-docstrings==1.3.0
flake8-polyfill==1.0.2
flake8-print==3.1.0
flake8-quotes==3.0.0
flake8-string-format==0.2.3
flake8-todo==0.7
flake8-annotations-complexity==0.0.6
flake8-variables-names==0.0.2
flake8-class-attributes-order==0.1.2
flake8-broken-line==0.1.0
mypy==1.10.0
pep8-naming==0.7.0
rozental-as-a-service==1.1.2
passlib==1.7.1
gunicorn==22.0.0
uvloop==0.19.0
click==7.0
apispec==3.0.0
freezegun==0.3.15
